requirements.txt
langchain
python-dotenv
groq
langchain_groq
langchain_huggingface
langchain_community
pypdf
chromadb
langchain-chroma
gradio
langchain_openai